// src/api/generated/models/TemplateMetadata.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { TemplateName } from "./TemplateName";
export type TemplateMetadata = {
  name: TemplateName;
  description: string;
  category: string;
  beta?: boolean;
  is_chronological?: boolean;
};